Move web/src/api to web/src/openapi
