+++ web/src/api-client.ts
@@ -1,3 +1,3 @@
-import { UsersApi } from "./api";
+import { UsersApi } from "./openapi";
 
 export const usersApi = new UsersApi();
