+++ web/src/openapi/.npmignore
