+++ web/src/openapi/.openapi-generator-ignore
