+++ web/src/openapi/.openapi-generator/VERSION
