+++ web/src/openapi/api.ts
