+++ web/src/openapi/common.ts
