+++ web/src/openapi/configuration.ts
